+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### v0.2.11 ???
-* Handle games without settings/hp_settings/equipment.bin
+* Changed call to skeleton generator to use subprocess.run instead of os.system, should capture stderr and stdout 
 * Better handling of GLTF2 export of adf files that are not models and/or are missing type info
+* Handle games without settings/hp_settings/equipment.bin
 * Regular expression mask:
     * Accepted when **Enter** is key'ed or **SET** is pressed
     * Reset to previously accepted value if **Esc** is key'ed or **CLEAR** is pressed